Test both sync and async gateway search
Gateway searching tests extracted out of async only code into a common
test module.
