+++ src/common/mod.rs
@@ -1,6 +1,8 @@
 pub mod messages;
 pub mod options;
 pub mod parsing;
+#[cfg(test)]
+mod tests;
 
 pub use self::options::SearchOptions;
